Technicaly done

+++ trunk/VisualProfilerUI/Model/CallTreeConvertors/Sampling/SamplingGlobalAggregatedValues.cs
@@ -2,7 +2,7 @@ namespace VisualProfilerUI.Model.CallTreeConvertors.Sampling
 {
     public class SamplingGlobalAggregatedValues
     {
-        public ulong WallClockDurationHns { get; set; }
+        public double WallClockDurationHns { get; set; }
         public ulong TotalActiveTime { get; set; }
         public uint StackTopOccurrenceCount { get; set; }
         public uint LastProfiledFrameInStackCount { get; set; }
